[BROKEN] Add layer between app & BlacklistDatabase
This class (blacklist.Blacklist) provides convenience methods for
interacting with the "lower-level" database layer.

Added another column to the BlacklistDatabase, 'make_permanent'. This
acts as a boolean (0 or 1) indicating that a blacklist record should be
made permanent when it expires. This exists to prevent users from
circumventing a temporary ban by requesting to be permanently
blacklisted and then requesting to be unblacklisted.

Renamed prefix.py -> redditprefix.py (the variable prefix was too
common).

Added a TODO block in src/bot.py@~515. I should have though about the
blacklisting messaging system a bit more before diving in because this
will be the second "refactor" just after writing the initial logic.
(Maybe refactor #1.5.)

The basic implementation idea I have now is:
        - lock the blacklist.Blacklist database calls (remove the
          database locks)
        - remove the blacklist_queue; let messages directly call
          blacklist.Blacklist methods
        - factor Reddit instantiation to its own method so that
          messages.run_forever can create its own (separate Reddit
          instances are process-safe). See:
https://praw.readthedocs.io/en/latest/getting_started/multiple_instances.html