[BROKEN] Remove locking from Database
I think it's better to keep Database relatively low-level and create an
intermediate interfacing level (eg. Blacklist <-> BlacklistDatabase) if
process-proofing is required for any given Database. This is especially
true for Blacklist because a lot of the Database methods are chained
together. If locking is implemented at the Database level, then any
cached state could become stale:
        eg.
                1. process A: check if userA is blacklisted
                2. process B: insert userA into blacklist
                3. process A: do stuff based on whether userA is
                        blacklisted

TODO: add locking to blacklist.Blacklist