[BROKEN] Refactor messages to use Blacklist
Instead of using a queue to communicate with the main process, the
messages sub-process will now directly modify the blacklist (since
blacklist handles locking).

Added a reply on successful blacklist add/remove.

Also added a debug pm to the AUTHOR account with debugging information
if something goes catastrohpically wrong (for now only sends this pm if
the subject regex matches but is not considered an add or remove -- ie,
the regex was altered to accept more subject strings but the code was
not updated to reflect).