[BROKEN] Remove ig user's db if missing fetch data
If the /media endpoint responds with an empty data['items'] array then
the user's profile is either private or otherwise has no public media
items. If we had already cached data for this user, then the previous
implementation would yield old media links if the user changed their
profile to private. This commit fixes the issue by deleting the user's
cached database file in this situation.