[BROKEN] Implemented StreamMixin
StreamMixin defines a stream property which can be used as a generator
for the child-defined stream method (eg. inbox.messages). This mixin
handles delaying retries in the event of an internet hiccup or server
error. Child classes only need to define the _stream_method property if
they rely on the default behavior of _stream.

Also refactored ProcessBase -> ProcessMixin.