[BROKEN?] Refactor reddit RATELIMIT handling
My original assumption was that seeing a RATELIMIT APIException meant
that ALL requests to reddit would receive the same error. It seems as
though that assumption was wrong (maybe).

My assumption now is that seeing a RATELIMIT means that POST methods are
rate-limited (that reddit streams are still fetchable). The original
implementation was simply sleeping all processes since I thought all
requests would be denied.

This commit refactors this system so that rate-limited praw requests
(ie, comment.reply, message.reply, etc) are now queued into a database
which are handled in a separate process. (The actual reddit rate-limit
timing is handled by another process which basically just sleeps,
waiting for the system to be rate-limited then sleeping until the
rate-limit is over and clearing the flag.)

This rate-limit handling replaces the previous error_handling module.