[BROKEN?] Write pids to file
Refactored run_forever to RunForeverMixin (which ProcessMixin now
inherits from). This was so that igHighlightsBot could inherit from it
without needing the extra process functionality.

All RunForeverMixin's will first check if a pidfile exists (and exit if
yes), otherwise it will write its pid to file and remove it when
run_forever exits.