[BROKEN?] Make submission queue persistent
If the program terminated unexpectedly, any mention-summoned submissions
in the queue would be lost. This should no longer be an issue with a
persistent queue.

This solution does bring another issue, however: if a submission is
fully processed and dies before it can commit the removal of the queue
element to disk, then the next run will result in that submission being
processed again.