[BROKEN?] Refactor replying out of main process
The motivation behind this refactor was that the queue processing in the
main process was dependent on breaks in the comment stream. So if the
set of subreddits grew too large or added too many popular subreddits,
then it was possible for the queues to never be processed.

This commit refactors all comment replies into a separate process
(replier.py). The main process (bot.py) and the mentions process
(mentions.py) enqueue comments that the bot should reply to and the
replier process formats and sends the replies.

The submission_queue was removed because it is no longer needed as its
functionality is now handled directly by the mentions process.