[BROKEN?] Handle duplicate badactor record
Previously a user could be flagged as a badactor multiple times for the
same thing (comment, submission, etc) if the previous flag had already
expired. This commit fixes this by checking both that 1) the thing is
not in either table of the database and 2) catching the
UniqueConstraintFailed if, somehow, a duplicate flag does occur.