Refactor reddit ratelimit event/value flag
Refactored the Event/Value into a single class, src.ratelimit.Flag. This
essentially just wraps the functionality of setting/clearing the Event
and Value.

It also provides a persistence for reddit ratelimiting so that the bot
does not have to hit reddit's servers to know that we are ratelimited.
This could happen if eg. a new comment is fetched that the bot wants to
reply to.