Fix _RateLimit process unresponsive exit
The _RateLimit process would sometimes become seemingly unresponsive to
kill signals because the rate_limited.wait() was not returning in a
timely fashion.

This was caused by the delay variable was being overwritten; there was
no real reason to use a variable for the wait() timeout anyway so it has
been removed.