Quiet reddit-ratelimit database log spam (e3d1068)
This was missed from e3d1068 because this flood of log output only
occurs when the bot is ratelimited by reddit. It was logging output
every timeout interval (= 1s as of this commit).

The __log_execute method was refactored to the base Database class so
that any derived class can now limit specific logging calls.