Comment out (temporarily?) mentions bad actors
There are too many false positives when considering mention summons as
bad behavior. The false positives by themselves aren't really an issue
since the temporary ban is only handed out once that user's bad-actor
count breaks the threshold. However, trolling the bot in general
probably won't be too much of an issue since there is no real feedback
that the offending user is actually doing anything to the bot.

I'm leaving 404 flagging on because too many of these probably is
indicative of some type of strange trolling. Or that user is just clumsy
with how they type.