Change to-add subreddit req to finding valid links
In order for summoned-to subreddits to be added, the requirement was the
bot posting at least one reply to the summoned submission. If the bot
is/was ratelimited then the reply would technically not count as
successful, messing up the to-add count.

This commit changes the requirement to only be the bot finding valid
instagram user links in the submission. Conceptually, this makes more
sense since the bot should probably be crawling subreddits where users
are regularly posting instagram user pages.