Fix bot sometimes posting duplicate instagram user
If the bot is summoned to a post that already has duplicate instagram
user links and becomes ratelimited then it would post a reply to each
instagram user since there was no check on whether that user was queued
to be posted.

This commit now prunes instagram users that have either been posted or
are currently queued to be posted in the reddit ratelimit queue.

Note: this implementation relies on the queued reply bodies to match the
Formatter's format otherwise it may not find any users. (This is only an
issue if the formatting changes significantly.)