Cache parsed comments for a bit
This adds a caching system for eached parsed comment. This should lower
the potential network hits of re-parsing comment bodies at the expense
of a somewhat higher memory footprint.

One issue with this implementation is that the clearing of the cache is
done in a very broad fashion. It simply clears the cache whenever it has
been too long since the last clearing of the cache. This may cause some
comments to be parsed twice if they occur on the straddling line between
about-to-clear and just-cleared.

This was motivated by wanting to quiet logging spam from the Replier
when an instagram user has been queued which would cause the Replier to
re-parse the same comment(s) repeatedly.