Refactor ig-queue to only track instagram users
In the situation where the bot queues 2+ unique comments for a reply
that both contain the same instagram user and that user becomes enqueued
into the instagram queue (eg. ratelimited), then the reply handling
would result in that user's data being fetched for every comment. The
expected behavior would be that the user's data is fetched for the first
comment and then the cached data is used for each comment thereafter.

This commit restructures the InstagramQueueDatabase so that it only
tracks instagram usernames and their corresponding last_id. Note: this
database is no longer really a queue in the FIFO sense (and it hasn't
been for some time now); it does hold data to be used later though, so
still sort of a queue.