Fix instagram ratelimit handling
Instagram ratelimited fetches were not continuing properly for two
reasons: 1) fetches were only triggered if the cache was expired (not
typically the case if the previous fetch was queued) and 2) last_ids
were not being updated if the instagram user was already queued.

Fixed the ratelimit time remaining output which was just showing the
timestamp (time.time()) before.