Fix interrupted instagram fetch over-pruning cache
The seen set was not persistent so the program's knowledge of which
items had been seen/not seen would not survive a program restart. This
commit fixes this issue by creating a "temporary" database that only
stores items it has encountered. When the fetch completes, it will prune
missing items from the actual database then remove the temporary
database. Note: temporary means the database should only exist until it
is no longer needed, not that the database is stored in /tmp or a tmpfs.

I chose to implement it this way instead of simply overwriting the old
cache so that, in theory, data is never lost. And in case I ever switch
on the stop_at_first_duplicate flag (switching it on during an
interrupted fetch would cause data loss).