Fix b31448c failed fetch from ig-queueing user
Since the top_media property is called twice on a failed fetch, the
second call would cause the instagram user to be queued if instagram was
ratelimited.

Refactored instagram ratelimited logging so that it only happens once
per state change (not ratelimited -> ratelimited & ratelimited -> not
ratelimited). Also added a check to prevent queueing invalid data for an
instagram user and re-queueing the same data.