Apply SUBREDDITS/BLACKLIST changes every startup
This commit changes the way the SubredditsDatabase and BlacklistDatabase
are initialized. Instead of only reading these default lists to
initialize their respective databases, they are now read every time the
program starts and the changes are applied to the databases. This way
the user has an easier way to add many subreddits.