Add -d, --dry-run
This option prevents the bot from replying to comments. It will still
process messages: add/remove blacklists and reply to messages. As well
as everything before replying to coments including parsing and
formatting the reply.

The motivation behind this was to release the bot into the wild without
it actually responding to comments so I can catch any edge-case bugs I
haven't considered. This is an imperfect solution to the problem since
it doesn't actually touch the reddit ratelimit code but it should expose
bugs related to fetching comments and instagram user data.