Increase Mentions/Messages run_forever delay
These don't really need to happen very frequently but probably shouldn't
be so long that someone requesting a blacklist needs to wait a long time
for a reply. To this end, 60s seems like a decent delay so that the log
isn't spammed/less network hits are made and a somewhat snappy response
time is maintained.