Refactor reddit.display_id to return permalink
The typical use-case for this method is to help determine what the thing
is. For comments (and maybe messages) this was not very useful since
there is no quick way to see the comment with just its id without
instantiating a praw.models.Comment object.

To correct this, display_id now returns the passed thing's permalink.
The method's name wasn't changed since the functionality is basically
the same.

Note: this does potentially add an extra network hit for each thing.

Also changed Replier's "Replying ..." logging to use display_id rather
than simply outputting the comment.id.