Don't re-raise Forbidden exception on reply
I believe encountering this exception means that the bot was banned from
the subreddit it was replying in. (It could also mean that the bot auth
credentials were incorrect but that shouldn't happen since the
constructor is checking that.)

So, encountering a Forbidden exception is no longer re-raised if
encountered on thing.reply().

Note: I am not yet adding the subreddit to the blacklist because I want
to verify that encountering this exception when replying actually means
the bot was banned.