Implement downvote-to-delete
This commit implements the downvote-to-delete feature. The bot will now
crawl its most controversial comments and delete ones whose scores have
dropped below a config-defined threshold.

The delay between comment score checks is scaled based on the lowest
encountered comment's score that is above the threshold such that the
process will wait a shorter period the closer to the threshold that
comment's score was. For example, seeing a comment score=-2 with
threshold=-5 will cause the bot to wait a shorter amount of time than
seeing a lowest comment score=100.

This was motivated by eb8f70d sometimes replying with incorrect
instagram user data.