Ignore soft-links from AutoModerator
Some subreddits have AutoModerator softlinking to a non-existant user as
an example of how to format posts (eg. '(@Instagram_user)'). This commit
does two things: 1) it fixes the soft-link username regex to include
cases where it is wrapped in brackets and 2) it ignores soft-linked
usernames posted by AutoModerator.

The regex implementation is just a hard-coded set of acceptable
delimiting characters. Is there a better way to do this so that I won't
have to manually update these sets every time a new acceptable
delimiting character is found?