Lower impact of outliers in sorting IG data
Relatively low like-count, high comment-count outliers were being
weighted too heavily. These outliers don't appear to be very indicative
of the instagram user's posts. (eg. sponsored giveaways generate a lot
of comments and a relatively low amount of likes.)

This commit fixes this by weighting the like-count inversely
proportional to the comment-count weight. So, as the comment:like ratio
approaches 1, the like-count weight will be high and the comment-count
weight will be low. As the comment:like ratio approaches 0, the
like-count weight will be low and the comment-count weight will be high.
ie,
        let L = num_likes, C = num_comments
        C / L -> 1, favor L
        C / L -> 0, favor C

In practice, this will largely sort the data by likes > comments unless
there are a very large amount of comments. The impact of the comments
weight doesn't appear to cause a dramatic shift in ordering with the
chosen constants. That is, a large comments outlier will only sort the
item higher up to a threshold based on that item's like-count. (At least
that's what seems to happen from my tests.)