Exclude outlier comment counts in IG data
This is related to 6e8a2f4 and 00c4589.

Outliers were still impacting sorting of some instagram caches. Others
were not properly reflecting their user's typical media.

This commit attempts to solve both by 1) excluding significant comment
count outliers (ones that are very, very far from the rest of the data)
and 2) ordering the data by likes scaled relative to the normalized
comments. This way, media that have relatively low comments are not
sorted as high even if they have a lot of likes.

Note: the outlier exclusion assumes that comment counts which exceed the
threshold are not indicative of the user's posts, that they are not just
extremely popular. (TODO? check like count before excluding; not
entirely sure how to include the like count though)