Exclude IG posts with high comments:likes
This is related to c828a29, ab84baf, and 19e1f61.

This commit accounts for media that has a very high comments-to-likes
ratio (by excluding it from the sort order). This can happen if the
post's number of comments does not exceed the outlier threshold (ie, it
is not an outlier). Posts with this property (high comments, low-ish
likes) are usually not very indicative of the user's posts.

The ratio threshold (0.08) is arbitrary (it seems to work for the caches
I tested with).