Sort IG data by likes if comments too low
The data was being sorted haphazardly if comments were too low (eg. 0).
This commit sorts low-comment posts by the like-count so that it isn't
as random (by capping the scale factor lower-bound to 0.1).

In practice, this shouldn't really impact any of the bot's behavior
since these posts will most likely not fall into the range of replied
highlights. Catching this case in the event that something changes seems
sensible, though.