Fix edge-case stale ig-queue items
It was theoretically possible for instagram users to become stale in the
InstagramQueueDatabase if a comment containing a queued instagram user
was deleted before that user could be fetched again. This would cause
the next fetch (possibly days/weeks/months later) to continue the
previous fetch from the last saved position resulting in the bot missing
newer posts (if there were any).

This commit solves the issue by resetting ig users' last_id column to
None (NULL) if the elapsed time since their INSERT/UPDATE timestamp
exceeds an arbitrary threshold (set to 1d at the moment). This reset is
applied on every 'in' (__contains__) and 'get_last_id_for' call.

Note: a new column, timestamp, was added to the database so any data
existing in a previous version of the database will be lost once this
version of the program is run.