Try non-english words if no usernames are found
This expands the Parser's functionality by trying single non-english
words as instagram usernames. This will dramatically increase the
frequency of 404s and false positive replies.

The motivation for this change is because not all users soft-link with a
prepended '@' sign. Perhaps this feature is a bit too ambitious.

To help combat the frequency of bad matches, I've made the username
regex pattern more strict. It should now match the range of valid
instagram usernames.

This commit adds a dependency on pyenchant.