Match IG media links w/ "taken-by=..." query data
Sometimes users will only link a single post from an instagram user but
will still include the '.../?taken-by={username}...' query. This commit
gives the bot the ability to spot these usernames and reply accordingly.
Note that this may cause the bot to reply to multiple usernames that
seemingly were not linked.

The Parser will not always log the usernames found in a comment (if
any). Some debug logging has been added to the Parser's link matching as
well.

The instagram regex tests have been moved to its own file instead of
attaching itself to the parser's tests.