Expand matched IG keyword prefixes (2ddeb96)
This expands possible instagram prefix keywords to include things like
'Her name on instagram is ...' and 'Their instagram: ...'.

This commit breaks the on_insta_rant ('... him on Insta ...') regex
matching. The Parser should prune this particular case because the word
preceding 'on Insta' is 'him' which is english. However, if someone says
something like '... I hate this geeeeezer on instagram ...' then the bot
may reply with @geeeeezer links (if that user has public posts). This
should be a pretty rare case though.