Clean up 500-level instagram status code logging
This was logging a "human-readable" time.time() which isn't very
meaningful. The logging will now display the clock-time expiration of
the fetch delay instead.

Note: this will still be spammy since the log spam originates at the
Replier level. However, the replier (at the moment) does not know
whether any given user needs to be fetched or is already cached. So
preventing the log from being spammed when instagram responds with a
500-level status code requires more code than expected. This happens so
infrequently that maybe it's ok that the log is spammed when it does
happen.