Parse submissions in addition to comments
Some users title their posts with the instagram username which leads to
the username never appearing in the comments. This commit gives the bot
the ability to parse submissions for instagram hard- or soft-links. It
does so by adding another process which parses the subreddits'
submissions stream.

Much of the Parser logic has been refactored to the internal
_ParserStrategy class. Its subclasses, _CommentParser and
_SubmissionParser, handle comment and submission parsing, respectively.
The Parser object itself is basically just for storage now.

The subreddit stream fetching logic has been refactored out of bot.py to
its own Mixin class so that the Submissions class does not duplicate
that code.

make_pickle.py has been refactored to try pickling submissions in
addition to comments. This change causes it to make a second ratelimit
hit if the given id is not a comment.

The changes to the reply database are mainly renaming 'comment' ->
'thing' so that it is clear that the database handles more than just
comments. Similarly, the changes to replies/filter.py are renaming
'comment' -> 'thing' where appropriate. The reply-queue now also handles
things instead of comments and has been refactored accordingly as well.
Note that this will cause previous versions of these databases to be
deleted and all associated data will be lost.

The logic for re-constructing praw.model objects from fullnames has been
moved from ratelimit.py -> reddit.py so that the Replier can make use of
it to account for the reply-queue now containing either comments or
submissions.