Fix? 8d6249e extra database columns crash
The query that was supposed to drop extra columns was incorrect. It was
attempting SELECT non-existant columns to INSERT into the updated table,
causing the program to crash. Further, the logic determining if extra
columns existed was not accounting for columns being replaced (ie,
the table both missing and having extra columns).

This commit should fix these issues. Note, though, that this hasn't
really been tested at all.