Sort IG data by likes only if comments too low
Users that don't have much comment activity were being sorted
questionably. This changes the sorting order to solely num_likes for
users with comment-count averages that are too high relative to their
max comment-count. Comment-count averages tend to skew low as comment
activity rises; thus, if an average is too high then that should
indicate that the user's comment activity is not a reliable factor in
gauging a post's popularity.