Fix StreamMixins not fetching elements after error
If a praw stream encounters a retryable exception (RequestException,
ResponseException, ServerError) then that process would become stuck
looping infinitely. This would happen because it is not possible to
restart a generator if it throws an exception, resulting in the .stream
generator looping infinitely.

This commit fixes this by restarting the generator upon encountering a
recoverable exception. The .stream property will now also halt if the
._killed flag is set. Note that restarting the generator will cause
elements to be re-parsed.