Make --dry-run more robust
The intended purpose of --dry-run is for live testing without actually
comitting replies to reddit. This commit should allow --dry-run to now
hit all parts of the code instead of most of it (some of the replying
logic and all of the ratelimit logic was not being run previously).

Why not write some integration tests instead? 1) I'm lazy, 2) I'm not
entirely sure how to properly write integration tests that actually
tests all parts of the code, and 3) I'm lazy.

reddit.py will now raise Forbidden and RATELIMIT APIExceptions when
running in dry-run mode. These should be thrown on the first and second
call to do_reply, respectively. The rate at which they are raised
thereafter is randomly chosen on an exponentially decaying scale so that
every call to do_reply has a lower and lower chance to raise.

The database wrappers have been slightly refactored to account for which
run-mode the bot is in (normal/dry-run). The databases will use
different files depending on the mode except for the Instagram*
databases since these are not outward facing. In particular,
InstagramRateLimitDatabase should never use different instances so that
the bot always has consistent knowledge of its current instagram
ratelimit state.

The program now stores a psuedo-global in the constants module,
"dry_run" so that the value does not need to be passed to essentially
every part of the code.