Fix Parser not missing IG suffix in strings
If the instagram keyword suffix appeared in the middle of the string,
the Parser would fail to recognize its existance. This commit changes
the HAS_IG_KEYWORD_REGEX test so that it sees suffixes appearing in the
middle of the string.

eg. 'That\'s foobar on insta. They're super duper cool!'