Fix fallback config values only returning strings
Some config options are assumed to be ints/floats. This was causing bad
config values to crash the current process. This issue should now be
fixed.

The fallback tests are not very good. They only test that a not-None
value is returned. TODO? test against the actual default bot.cfg file?