Handle reddit ratelimit looping on Forbidden
If a thing was reddit ratelimit-queued where thing.reply() would raise a
Forbidden exception (the bot was either banned from the subreddit or
reddit), then the RateLimitHandler would infinitely retry the reply for
that thing. This situation could happen if the bot attempted to reply to
a Forbidden thing while already ratelimited (resulting in immediately
ratelimit-queueing the thing, skipping the .reply).

This commit resolves this issue by removing the thing from the ratelimit
queue on either reply success or reply not-possible. One side effect of
this implementation is that the bot will retry replies to Forbidden
things if it encounters them again. This side effect is not new. Also,
note that the bot currently has no concept of whether it has been banned
from a subreddit meaning it will try to reply to any reply-able thing in
subreddits that it has been banned from.