Reload the config if its mtime changes
The config will now lazy-reload from file if it has been modified (mtime
changed). Lazy-reload means that a reload will only happen if one of the
config object's properties is accessed. Note that each process will
reload the config if it has changed.

The following options will not change the program's behavior if modified
while it is running:
        - praw_sitename
        - app_name
        - logging_path
        - logging_level
        - colorful_logs
The first two would introduce much more complexity to the config
reloading logic. The logging_level and colorful_logs options would not
be terribly difficult to implement; I'm not sure about logging_path.