Ignore AutoModerator completely (d85ead1)
I think the threshold to show replies to AutoMod comments is higher so
all comments/posts by AutoModerator are now ignored by the bot. That is,
replies to AutoMod are hidden behind '(N more comments)' by default.

Some subreddits have AutoModerator posting links to instagram users
which it gets from the submission title. The bot is already parsing
these titles. This was causing a sort of race condition between the bot
replying to submissions and AutoMod comments.