Fix the bot sometimes making duplicate replies
The bot would sometimes reply with the same instagram account multiple
times in a submission. It would see that it had already posted the
username but it would not be pruned. This was happening because the
parser-found username would maintain the comment author's capitalization
whereas the replies-database would ignore casing, resulting in the
username not being pruned and ultimately the bot making a duplicate
reply.