Fix Parser sometimes not matching consecutive @s
The Parser would sometimes miss soft-linked @usernames if they appeared
consecutively.
eg.
        '@foo @bar @baz' would only match ['foo', 'baz']
This was happening because the the '\s+' was being matched for the first
and last usernames resulting in the middle to have no spaces to match
against.

This should be resolved but may cause some overmatching, though all the
current tests pass so maybe not.