Fix @username overmatching (84bf20a)
The bot was sometimes partially matching invalid, non-ascii usernames
(eg. @Alíca.dávis). This was happening due to the change from 84bf20a
to accomodate consecutive @usernames.

This commit enforces one of: an ending delimiting character, one or more
spaces, or the end of the string in order to match an @username. The 1+
spaces was only an optional requirement previously which was causing
usernames beginning with valid characters to partially match
(@Alíca.dávis -> @Al).