Cull outlier comments more aggressively
This prunes outlier comment-count instagram posts more aggressively by
constricting the minimum normalized like-count percentage to 75% instead
of 25%.

The commit includes a median implementation because I was playing around
with normalizing the like-count relative to the median rather than the
minimum. I don't think there is really any point to doing so though.