Sort IG data by likes if comment activity too low
This is related to 8c9fca0.

Some users data sorts strangely if their comment activity is too low.
This commit attempts to correct the sort order by sorting the data by
num_likes if the avg comment-count is too low. The chosen too-low value
is arbitrary but appears to work with the sparse tests I did.