Lower? logger color memory footprint
Every message that was colorized by the logger was being held onto even
if the message was only displayed a single time. This commit removes
colorized entries that the logger has not seen in a 3-hour interval in
an attempt to lower the logger's memory footprint.

This change will cause infrequently used colorized messages to change
colors.