Link non-linked/guessed private profile pages
With this commit, the bot will now reply to non-linked, non-guessed
private profile pages. This can be useful if an instagram account is
only private to control comments/reports.

The /?__a=1 endpoint is now hit at the beginning of each new fetch to
determine if the account exists and is private. This adds an extra
ratelimit hit for each fetch but that should remain negligible. This
change does however render the flag_as_bad logic at the end of __fetch
obsolete. It has not been removed because I plan to refactor the
instagram module. On a related note, a lot of the new instagram logic
has been duplicated (copy/pasted) for the same reason.

Instagram databases now differentiate between BAD (non-existant user)
and PRIVATE (private account) using the same method as the "bad"
flagging.

The src/replies changes are mainly to facilitate private instagram
account linking. Note: that I've only performed rudimentary tests on the
Formatter.  There may be bugs related to multi-reply formatting. (The
bot has yet to commit more than a single reply -- 10k chars is quite a
lot.)