Skip low follower instagram accounts
The bot will no longer reply with links/highlights to instagram accounts
that do not meet a minimum threshold of followers (1000 by default).
This should hopefully help in preventing the bot from inadvertently
doxxing instagram users. It also has the added benefit of pruning some
false positive username guesses.

Not-enough-follower usernames are being flagged as bad so that repeated
fetches are not needlessly attempted. If a popular user recreates their
account for whatever reason and rapidly grows in follower count then
flagging the account as bad will cause no replies for the new account
over the expiration period (whereas not flagging the username may cause
the bot to reply if the follower count surpasses the threshold). This is
probably an extremely rare case. The number of ratelimit hits prevented
probably outweighs the number of accounts that would have surpassed the
minimum follower threshold. In either case, in terms of the ratelimit,
this logic is negligible.