Fix instagram fetch: /media shutdown
As of Nov 7, 2017, the /media endpoint always returns 404. There is no
telling if the /?__a=1 endpoint will also be shutdown but it is the only
other non-api endpoint I know of. If this endpoint is shutdown, the bot
will have to scrape instagram's actual site to collect data.

This is a hotfix just to get the bot working again. There is an
extraneous hit from get_meta_data since fetch_data now hits the same
endpoint. I intend to fix this.