Sleep Replier, RateLimitHandler if ratelimited
Neither of these processes were sleeping if the bot was ratelimited from
posting to reddit. This wasn't actually causing any ill side effects but
there was no reason for the processes to be active in this case.

Reddit's post ratelimiting really only applies to very new (ie, very low
karma) accounts. The bot has not actually hit the posting ratelimit
since the first day it was live.