Fix SubredditsDatabase wrongly seeding from file
The default behavior for the SubredditsDatabase is to seed from the
SUBREDDITS file if the database file does not exist. This was always
evaluating to True because the .PATH constant is only specifying the
database filename (basename).

This has been a bug since v0.8 (a073bbb); ie, since the .PATH constant
was changed.