Lower Replier spam when instagram ratelimited
This gives the Replier two in-memory caches (dictionaries) to store 1.
reply-queued things (comments/submissions) and 2. things that require
instagram fetches. The first allows the replier to skip duplicated
reddit network hits for re-parsed things. The second allows the replier
to skip things that require instagram fetches when the bot is instagram
ratelimited.

The main motivation behind this change was to lower logging spam when
instagram ratelimited. This implementation has the added benefit of
lowering unnecessary network hits.