Turn off self-imposed IG database ratelimiting
The InstagramRateLimitDatabase still exists but is no longer used as of
this commit (similar to the BadActorsDatabase).

The database ratelimiting is no longer needed because instagram responds
with a retry delay time when they ratelimit the bot with 429 Too Many
Requests. That is, the bot no longer needs to track and guess
instagram's ratelimiting rules since they tell us explicitly (well, the
timing is explicit anyway).