Fix removal of dynamically added subs on bootup
This bug has existed since 955a82b.

Dynamically added subreddits (eg. from potential-subreddits) were being
removed if they were not defined in the SUBREDDITS file on bootup. This
commit fixes this by no longer removing subreddits that are not present
in the SUBREDDITS file on startup. The downside to this implementation
is that the SUBREDDITS file is no longer a direct 1:1 match with the
subreddits database.