Allow Parser to guess markdown quoted strings
This gives the bot the ability to guess single-word markdown quoted
strings as usernames.

The observed use-case for this is if the original user mentions a
username but includes bad characters before or after and another user
quotes the first but cleans up the string so that it is only the
username.

eg.
user A: 'Karmabirdfly....shes got a Tumblr and Instagram'
user B: '> Karmabirdfly' [...]