Ignore the bot's FAQ post
Because of the Submitter process, the bot now crawls its own profile.
The FAQ post contains an instagram link at the moment and may contain
more in the future.

This commit gives the bot the ability to ignore specific things by id.
This implementation is probably a bit too generic for the problem but
opens the option of ignoring other things in the future.