Fix Submitter rapid-fire posting when ratelimited
If the do_submit call returned False (eg. when reddit ratelimited), then
the Submitter process would choose & attempt to post a new instagram
user every iteration. This commit fixes the issue by using a
non-boolean, non-None value as the "not-set" value so that the inner
posting loop will exit if any do_submit attempt is made.