Fix RedditRateLimitQueueDatabase not deleting rows
This was an issue introduced in af6d172. SQLite treats NULLs as distinct
values for UNIQUE constraints and WHERE clauses. This was causing
strange behavior when dealing with the database such as rows not being
deleted.

This commit resolves the issue by replacing NULL (None) with a
placeholder value that SQLite should treat as a non-distinct value.

See https://www.sqlite.org/nulls.html for more information about how
SQLite handles NULLs.